cloud_functions/fetch_weather_data_function/requirements.txt
google-cloud-bigquery
google-cloud-storage
google-cloud-pubsub
openmeteo-requests
requests-cache
retry-requests